backend/requirements.txt
fastapi>=0.100.0
pydantic>=2.4.1,<2.12.0
pydantic[email]>=2.4.1,<2.12.0
uvicorn>=0.15.0
sqlalchemy>=1.4.0
alembic>=1.7.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.5
emails>=0.6.0
aiogram>=3.3.0,<3.4.0
python-dotenv>=0.19.0
fastapi-mail>=1.4.1
aiosmtplib>=2.0.2
aiofiles>=23.2.1
PyJWT>=2.8.0
asyncio
httpx
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
ruff==0.6.9
black==24.8.0
isort==5.13.2
Pillow>=10.0.0
boto3>=1.34.0
botocore>=1.34.0
